@@ -70,10 +70,12 @@ DIALECTS = [
     "ClickHouse",
     "Databricks",
     "Doris",
+    "Dremio",
     "Drill",
     "Druid",
     "DuckDB",
     "Dune",
+    "Fabric",
     "Hive",
     "Materialize",
     "MySQL",
@@ -83,6 +85,7 @@ DIALECTS = [
     "PRQL",
     "Redshift",
     "RisingWave",
+    "SingleStore",
     "Snowflake",
     "Spark",
     "Spark2",
@@ -92,6 +95,7 @@ DIALECTS = [
     "Teradata",
     "Trino",
     "TSQL",
+    "Exasol",
     "E6",
 ]
 
@@ -106,7 +110,10 @@ MODULE_BY_ATTRIBUTE = {
 
 __all__ = list(MODULE_BY_ATTRIBUTE)
 
-_import_lock = threading.Lock()
+# We use a reentrant lock because a dialect may depend on (i.e., import) other dialects.
+# Without it, the first dialect import would never be completed, because subsequent
+# imports would be blocked on the lock held by the first import.
+_import_lock = threading.RLock()
 
 
 def __getattr__(name):
