@@ -41,7 +41,7 @@ def parse(path: str, dialect: DialectType = None) -> exp.JSONPath:
     """Takes in a JSON path string and parses it into a JSONPath expression."""
     from sqlglot.dialects import Dialect
 
-    jsonpath_tokenizer = Dialect.get_or_raise(dialect).jsonpath_tokenizer
+    jsonpath_tokenizer = Dialect.get_or_raise(dialect).jsonpath_tokenizer()
     tokens = jsonpath_tokenizer.tokenize(path)
     size = len(tokens)
 
