@@ -358,13 +358,20 @@ def unnest_to_explode(
             has_multi_expr = len(exprs) > 1
             this, *expressions = _unnest_zip_exprs(unnest, exprs, has_multi_expr)
 
+            columns = alias.columns if alias else []
+            offset = unnest.args.get("offset")
+            if offset:
+                columns.insert(
+                    0, offset if isinstance(offset, exp.Identifier) else exp.to_identifier("pos")
+                )
+
             unnest.replace(
                 exp.Table(
                     this=_udtf_type(unnest, has_multi_expr)(
                         this=this,
                         expressions=expressions,
                     ),
-                    alias=exp.TableAlias(this=alias.this, columns=alias.columns) if alias else None,
+                    alias=exp.TableAlias(this=alias.this, columns=columns) if alias else None,
                 )
             )
 
@@ -397,6 +404,13 @@ def unnest_to_explode(
                 if not has_multi_expr and len(alias_cols) not in (1, 2):
                     raise UnsupportedError(
                         "CROSS JOIN UNNEST to LATERAL VIEW EXPLODE transformation requires explicit column aliases"
+                    )
+
+                offset = unnest.args.get("offset")
+                if offset:
+                    alias_cols.insert(
+                        0,
+                        offset if isinstance(offset, exp.Identifier) else exp.to_identifier("pos"),
                     )
 
                 for e, column in zip(exprs, alias_cols):
@@ -852,113 +866,122 @@ def struct_kv_to_alias(expression: exp.Expression) -> exp.Expression:
 
 
 def eliminate_join_marks(expression: exp.Expression) -> exp.Expression:
+    """https://docs.oracle.com/cd/B19306_01/server.102/b14200/queries006.htm#sthref3178
+
+    1. You cannot specify the (+) operator in a query block that also contains FROM clause join syntax.
+
+    2. The (+) operator can appear only in the WHERE clause or, in the context of left-correlation (that is, when specifying the TABLE clause) in the FROM clause, and can be applied only to a column of a table or view.
+
+    The (+) operator does not produce an outer join if you specify one table in the outer query and the other table in an inner query.
+
+    You cannot use the (+) operator to outer-join a table to itself, although self joins are valid.
+
+    The (+) operator can be applied only to a column, not to an arbitrary expression. However, an arbitrary expression can contain one or more columns marked with the (+) operator.
+
+    A WHERE condition containing the (+) operator cannot be combined with another condition using the OR logical operator.
+
+    A WHERE condition cannot use the IN comparison condition to compare a column marked with the (+) operator with an expression.
+
+    A WHERE condition cannot compare any column marked with the (+) operator with a subquery.
+
+    -- example with WHERE
+    SELECT d.department_name, sum(e.salary) as total_salary
+    FROM departments d, employees e
+    WHERE e.department_id(+) = d.department_id
+    group by department_name
+
+    -- example of left correlation in select
+    SELECT d.department_name, (
+        SELECT SUM(e.salary)
+            FROM employees e
+            WHERE e.department_id(+) = d.department_id) AS total_salary
+    FROM departments d;
+
+    -- example of left correlation in from
+    SELECT d.department_name, t.total_salary
+    FROM departments d, (
+            SELECT SUM(e.salary) AS total_salary
+            FROM employees e
+            WHERE e.department_id(+) = d.department_id
+        ) t
     """
-    Remove join marks from an AST. This rule assumes that all marked columns are qualified.
-    If this does not hold for a query, consider running `sqlglot.optimizer.qualify` first.
 
-    For example,
-        SELECT * FROM a, b WHERE a.id = b.id(+)    -- ... is converted to
-        SELECT * FROM a LEFT JOIN b ON a.id = b.id -- this
-
-    Args:
-        expression: The AST to remove join marks from.
-
-    Returns:
-       The AST with join marks removed.
-    """
     from sqlglot.optimizer.scope import traverse_scope
+    from sqlglot.optimizer.normalize import normalize, normalized
+    from collections import defaultdict
 
-    for scope in traverse_scope(expression):
+    # we go in reverse to check the main query for left correlation
+    for scope in reversed(traverse_scope(expression)):
         query = scope.expression
 
         where = query.args.get("where")
-        joins = query.args.get("joins")
+        joins = query.args.get("joins", [])
 
-        if not where or not joins:
+        # knockout: we do not support left correlation (see point 2)
+        assert not scope.is_correlated_subquery, "Correlated queries are not supported"
+
+        # nothing to do - we check it here after knockout above
+        if not where or not any(c.args.get("join_mark") for c in where.find_all(exp.Column)):
             continue
 
-        query_from = query.args["from"]
+        # make sure we have AND of ORs to have clear join terms
+        where = normalize(where.this)
+        assert normalized(where), "Cannot normalize JOIN predicates"
 
-        # These keep track of the joins to be replaced
-        new_joins: t.Dict[str, exp.Join] = {}
-        old_joins = {join.alias_or_name: join for join in joins}
+        joins_ons = defaultdict(list)  # dict of {name: list of join AND conditions}
+        for cond in [where] if not isinstance(where, exp.And) else where.flatten():
+            join_cols = [col for col in cond.find_all(exp.Column) if col.args.get("join_mark")]
 
-        for column in scope.columns:
-            if not column.args.get("join_mark"):
+            left_join_table = set(col.table for col in join_cols)
+            if not left_join_table:
                 continue
 
-            predicate = column.find_ancestor(exp.Predicate, exp.Select)
-            assert isinstance(
-                predicate, exp.Binary
-            ), "Columns can only be marked with (+) when involved in a binary operation"
-
-            predicate_parent = predicate.parent
-            join_predicate = predicate.pop()
-
-            left_columns = [
-                c for c in join_predicate.left.find_all(exp.Column) if c.args.get("join_mark")
-            ]
-            right_columns = [
-                c for c in join_predicate.right.find_all(exp.Column) if c.args.get("join_mark")
-            ]
-
             assert not (
-                left_columns and right_columns
-            ), "The (+) marker cannot appear in both sides of a binary predicate"
+                len(left_join_table) > 1
+            ), "Cannot combine JOIN predicates from different tables"
 
-            marked_column_tables = set()
-            for col in left_columns or right_columns:
-                table = col.table
-                assert table, f"Column {col} needs to be qualified with a table"
-
+            for col in join_cols:
                 col.set("join_mark", False)
-                marked_column_tables.add(table)
 
-            assert (
-                len(marked_column_tables) == 1
-            ), "Columns of only a single table can be marked with (+) in a given binary predicate"
+            joins_ons[left_join_table.pop()].append(cond)
 
-            # Add predicate if join already copied, or add join if it is new
-            join_this = old_joins.get(col.table, query_from).this
-            existing_join = new_joins.get(join_this.alias_or_name)
-            if existing_join:
-                existing_join.set("on", exp.and_(existing_join.args["on"], join_predicate))
-            else:
-                new_joins[join_this.alias_or_name] = exp.Join(
-                    this=join_this.copy(), on=join_predicate.copy(), kind="LEFT"
-                )
+        old_joins = {join.alias_or_name: join for join in joins}
+        new_joins = {}
+        query_from = query.args["from"]
 
-            # If the parent of the target predicate is a binary node, then it now has only one child
-            if isinstance(predicate_parent, exp.Binary):
-                if predicate_parent.left is None:
-                    predicate_parent.replace(predicate_parent.right)
-                else:
-                    predicate_parent.replace(predicate_parent.left)
+        for table, predicates in joins_ons.items():
+            join_what = old_joins.get(table, query_from).this.copy()
+            new_joins[join_what.alias_or_name] = exp.Join(
+                this=join_what, on=exp.and_(*predicates), kind="LEFT"
+            )
 
-        only_old_join_sources = old_joins.keys() - new_joins.keys()
+            for p in predicates:
+                while isinstance(p.parent, exp.Paren):
+                    p.parent.replace(p)
+
+                parent = p.parent
+                p.pop()
+                if isinstance(parent, exp.Binary):
+                    parent.replace(parent.right if parent.left is None else parent.left)
+                elif isinstance(parent, exp.Where):
+                    parent.pop()
 
         if query_from.alias_or_name in new_joins:
+            only_old_joins = old_joins.keys() - new_joins.keys()
             assert (
-                len(only_old_join_sources) >= 1
+                len(only_old_joins) >= 1
             ), "Cannot determine which table to use in the new FROM clause"
 
-            new_from_name = list(only_old_join_sources)[0]
-            query.set("from", exp.From(this=old_joins.pop(new_from_name).this))
-            only_old_join_sources.remove(new_from_name)
+            new_from_name = list(only_old_joins)[0]
+            query.set("from", exp.From(this=old_joins[new_from_name].this))
 
         if new_joins:
-            only_old_join_expressions = []
-            for old_join_source in only_old_join_sources:
-                old_join_expression = old_joins[old_join_source]
-                if not old_join_expression.kind:
-                    old_join_expression.set("kind", "CROSS")
-
-                only_old_join_expressions.append(old_join_expression)
-
-            query.set("joins", list(new_joins.values()) + only_old_join_expressions)
-
-        if not where.this:
-            where.pop()
+            for n, j in old_joins.items():  # preserve any other joins
+                if n not in new_joins and n != query.args["from"].name:
+                    if not j.kind:
+                        j.set("kind", "CROSS")
+                    new_joins[n] = j
+            query.set("joins", list(new_joins.values()))
 
     return expression
 
@@ -977,7 +1000,7 @@ def any_to_exists(expression: exp.Expression) -> exp.Expression:
     if isinstance(expression, exp.Select):
         for any_expr in expression.find_all(exp.Any):
             this = any_expr.this
-            if isinstance(this, exp.Query):
+            if isinstance(this, exp.Query) or isinstance(any_expr.parent, (exp.Like, exp.ILike)):
                 continue
 
             binop = any_expr.parent
