@@ -11,6 +11,7 @@ from sqlglot import (
     parse_one,
 )
 from sqlglot.dialects import BigQuery, Hive, Snowflake
+from sqlglot.dialects.dialect import Version
 from sqlglot.parser import logger as parser_logger
 
 
@@ -142,12 +143,23 @@ class TestDialect(Validator):
             "oracle, normalization_strategy = lowercase, version = 19.5"
         )
         self.assertEqual(oracle_with_settings.normalization_strategy.value, "LOWERCASE")
-        self.assertEqual(oracle_with_settings.settings, {"version": "19.5"})
+        self.assertEqual(oracle_with_settings.version, Version("19.5"))
 
-        bool_settings = Dialect.get_or_raise("oracle, s1=TruE, s2=1, s3=FaLse, s4=0, s5=nonbool")
+        class MyDialect(Dialect):
+            SUPPORTED_SETTINGS = {"s1", "s2", "s3", "s4", "s5"}
+
+        bool_settings = Dialect.get_or_raise("mydialect, s1=TruE, s2=1, s3=FaLse, s4=0, s5=nonbool")
         self.assertEqual(
             bool_settings.settings,
             {"s1": True, "s2": True, "s3": False, "s4": False, "s5": "nonbool"},
+        )
+
+        with self.assertRaises(ValueError) as cm:
+            Dialect.get_or_raise("tsql,normalisation_strategy=case_sensitive")
+
+        self.assertEqual(
+            "Unknown setting 'normalisation_strategy'. Did you mean normalization_strategy?",
+            str(cm.exception),
         )
 
     def test_compare_dialects(self):
@@ -178,7 +190,9 @@ class TestDialect(Validator):
 
     def test_compare_dialect_versions(self):
         ddb_v1 = Dialect.get_or_raise("duckdb, version=1.0")
-        ddb_v1_2 = Dialect.get_or_raise("duckdb, foo=bar, version=1.0")
+        ddb_v1_2 = Dialect.get_or_raise(
+            "duckdb, normalization_strategy=case_sensitive, version=1.0"
+        )
         ddb_v2 = Dialect.get_or_raise("duckdb, version=2.2.4")
         ddb_latest = Dialect.get_or_raise("duckdb")
 
@@ -456,6 +470,16 @@ class TestDialect(Validator):
             "CAST('127.0.0.1/32' AS INET)",
             read={"postgres": "INET '127.0.0.1/32'"},
         )
+        self.assertIsNotNone(
+            self.validate_identity("CREATE TABLE foo (bar INT AS (foo))").find(
+                exp.ComputedColumnConstraint
+            )
+        )
+        self.assertIsNotNone(
+            self.validate_identity(
+                "CREATE TABLE foo (t1 INT, t2 INT, bar INT AS (t1 * t2 * 2))"
+            ).find(exp.ComputedColumnConstraint)
+        )
 
     def test_ddl(self):
         self.validate_all(
@@ -526,51 +550,56 @@ class TestDialect(Validator):
         self.validate_all(
             "SELECT DECODE(a, 1, 'one')",
             write={
-                "": "SELECT CASE WHEN a = 1 THEN 'one' END",
-                "oracle": "SELECT CASE WHEN a = 1 THEN 'one' END",
-                "redshift": "SELECT CASE WHEN a = 1 THEN 'one' END",
-                "snowflake": "SELECT CASE WHEN a = 1 THEN 'one' END",
-                "spark": "SELECT CASE WHEN a = 1 THEN 'one' END",
+                "": "SELECT DECODE(a, 1, 'one')",
+                "duckdb": "SELECT CASE WHEN a = 1 THEN 'one' END",
+                "oracle": "SELECT DECODE(a, 1, 'one')",
+                "redshift": "SELECT DECODE(a, 1, 'one')",
+                "snowflake": "SELECT DECODE(a, 1, 'one')",
+                "spark": "SELECT DECODE(a, 1, 'one')",
             },
         )
         self.validate_all(
             "SELECT DECODE(a, 1, 'one', 'default')",
             write={
-                "": "SELECT CASE WHEN a = 1 THEN 'one' ELSE 'default' END",
-                "oracle": "SELECT CASE WHEN a = 1 THEN 'one' ELSE 'default' END",
-                "redshift": "SELECT CASE WHEN a = 1 THEN 'one' ELSE 'default' END",
-                "snowflake": "SELECT CASE WHEN a = 1 THEN 'one' ELSE 'default' END",
-                "spark": "SELECT CASE WHEN a = 1 THEN 'one' ELSE 'default' END",
+                "": "SELECT DECODE(a, 1, 'one', 'default')",
+                "duckdb": "SELECT CASE WHEN a = 1 THEN 'one' ELSE 'default' END",
+                "oracle": "SELECT DECODE(a, 1, 'one', 'default')",
+                "redshift": "SELECT DECODE(a, 1, 'one', 'default')",
+                "snowflake": "SELECT DECODE(a, 1, 'one', 'default')",
+                "spark": "SELECT DECODE(a, 1, 'one', 'default')",
             },
         )
         self.validate_all(
             "SELECT DECODE(a, NULL, 'null')",
             write={
-                "": "SELECT CASE WHEN a IS NULL THEN 'null' END",
-                "oracle": "SELECT CASE WHEN a IS NULL THEN 'null' END",
-                "redshift": "SELECT CASE WHEN a IS NULL THEN 'null' END",
-                "snowflake": "SELECT CASE WHEN a IS NULL THEN 'null' END",
-                "spark": "SELECT CASE WHEN a IS NULL THEN 'null' END",
+                "": "SELECT DECODE(a, NULL, 'null')",
+                "duckdb": "SELECT CASE WHEN a IS NULL THEN 'null' END",
+                "oracle": "SELECT DECODE(a, NULL, 'null')",
+                "redshift": "SELECT DECODE(a, NULL, 'null')",
+                "snowflake": "SELECT DECODE(a, NULL, 'null')",
+                "spark": "SELECT DECODE(a, NULL, 'null')",
             },
         )
         self.validate_all(
             "SELECT DECODE(a, b, c)",
             write={
-                "": "SELECT CASE WHEN a = b OR (a IS NULL AND b IS NULL) THEN c END",
-                "oracle": "SELECT CASE WHEN a = b OR (a IS NULL AND b IS NULL) THEN c END",
-                "redshift": "SELECT CASE WHEN a = b OR (a IS NULL AND b IS NULL) THEN c END",
-                "snowflake": "SELECT CASE WHEN a = b OR (a IS NULL AND b IS NULL) THEN c END",
-                "spark": "SELECT CASE WHEN a = b OR (a IS NULL AND b IS NULL) THEN c END",
+                "": "SELECT DECODE(a, b, c)",
+                "duckdb": "SELECT CASE WHEN a = b OR (a IS NULL AND b IS NULL) THEN c END",
+                "oracle": "SELECT DECODE(a, b, c)",
+                "redshift": "SELECT DECODE(a, b, c)",
+                "snowflake": "SELECT DECODE(a, b, c)",
+                "spark": "SELECT DECODE(a, b, c)",
             },
         )
         self.validate_all(
             "SELECT DECODE(tbl.col, 'some_string', 'foo')",
             write={
-                "": "SELECT CASE WHEN tbl.col = 'some_string' THEN 'foo' END",
-                "oracle": "SELECT CASE WHEN tbl.col = 'some_string' THEN 'foo' END",
-                "redshift": "SELECT CASE WHEN tbl.col = 'some_string' THEN 'foo' END",
-                "snowflake": "SELECT CASE WHEN tbl.col = 'some_string' THEN 'foo' END",
-                "spark": "SELECT CASE WHEN tbl.col = 'some_string' THEN 'foo' END",
+                "": "SELECT DECODE(tbl.col, 'some_string', 'foo')",
+                "duckdb": "SELECT CASE WHEN tbl.col = 'some_string' THEN 'foo' END",
+                "oracle": "SELECT DECODE(tbl.col, 'some_string', 'foo')",
+                "redshift": "SELECT DECODE(tbl.col, 'some_string', 'foo')",
+                "snowflake": "SELECT DECODE(tbl.col, 'some_string', 'foo')",
+                "spark": "SELECT DECODE(tbl.col, 'some_string', 'foo')",
             },
         )
 
@@ -611,6 +640,7 @@ class TestDialect(Validator):
                 "clickhouse": "SELECT CASE WHEN NOT (a IS NULL) THEN b ELSE c END",
                 "databricks": "SELECT NVL2(a, b, c)",
                 "doris": "SELECT CASE WHEN NOT a IS NULL THEN b ELSE c END",
+                "dremio": "SELECT CASE WHEN NOT a IS NULL THEN b ELSE c END",
                 "drill": "SELECT CASE WHEN NOT a IS NULL THEN b ELSE c END",
                 "duckdb": "SELECT CASE WHEN NOT a IS NULL THEN b ELSE c END",
                 "hive": "SELECT CASE WHEN NOT a IS NULL THEN b ELSE c END",
@@ -637,6 +667,7 @@ class TestDialect(Validator):
                 "clickhouse": "SELECT CASE WHEN NOT (a IS NULL) THEN b END",
                 "databricks": "SELECT NVL2(a, b)",
                 "doris": "SELECT CASE WHEN NOT a IS NULL THEN b END",
+                "dremio": "SELECT CASE WHEN NOT a IS NULL THEN b END",
                 "drill": "SELECT CASE WHEN NOT a IS NULL THEN b END",
                 "duckdb": "SELECT CASE WHEN NOT a IS NULL THEN b END",
                 "hive": "SELECT CASE WHEN NOT a IS NULL THEN b END",
@@ -955,6 +986,7 @@ class TestDialect(Validator):
             "DATE_ADD(x, 1, 'DAY')",
             read={
                 "snowflake": "DATEADD('DAY', 1, x)",
+                "dremio": "DATE_ADD(x, 1)",
             },
             write={
                 "bigquery": "DATE_ADD(x, INTERVAL 1 DAY)",
@@ -971,6 +1003,7 @@ class TestDialect(Validator):
                 "starrocks": "DATE_ADD(x, INTERVAL 1 DAY)",
                 "tsql": "DATEADD(DAY, 1, x)",
                 "doris": "DATE_ADD(x, INTERVAL 1 DAY)",
+                "dremio": "DATE_ADD(x, 1)",
             },
         )
         self.validate_all(
@@ -985,6 +1018,7 @@ class TestDialect(Validator):
                 "spark": "DATE_ADD(x, 1)",
                 "starrocks": "DATE_ADD(x, INTERVAL 1 DAY)",
                 "doris": "DATE_ADD(x, INTERVAL 1 DAY)",
+                "dremio": "DATE_ADD(x, 1)",
             },
         )
         self.validate_all(
@@ -1173,6 +1207,7 @@ class TestDialect(Validator):
                 "hive": "DATE_ADD(CAST('2020-01-01' AS DATE), 1)",
                 "presto": "DATE_ADD('DAY', 1, CAST('2020-01-01' AS DATE))",
                 "spark": "DATE_ADD(CAST('2020-01-01' AS DATE), 1)",
+                "dremio": "DATE_ADD(CAST('2020-01-01' AS DATE), 1)",
             },
         )
         self.validate_all(
@@ -1298,6 +1333,74 @@ class TestDialect(Validator):
             },
         )
 
+        self.validate_all(
+            "ARRAY_INTERSECT(x, y)",
+            read={
+                "hive": "ARRAY_INTERSECT(x, y)",
+                "spark2": "ARRAY_INTERSECT(x, y)",
+                "spark": "ARRAY_INTERSECT(x, y)",
+                "databricks": "ARRAY_INTERSECT(x, y)",
+                "presto": "ARRAY_INTERSECT(x, y)",
+                "trino": "ARRAY_INTERSECT(x, y)",
+                "snowflake": "ARRAY_INTERSECTION(x, y)",
+                "starrocks": "ARRAY_INTERSECT(x, y)",
+            },
+            write={
+                "hive": "ARRAY_INTERSECT(x, y)",
+                "spark2": "ARRAY_INTERSECT(x, y)",
+                "spark": "ARRAY_INTERSECT(x, y)",
+                "databricks": "ARRAY_INTERSECT(x, y)",
+                "presto": "ARRAY_INTERSECT(x, y)",
+                "trino": "ARRAY_INTERSECT(x, y)",
+                "snowflake": "ARRAY_INTERSECTION(x, y)",
+                "starrocks": "ARRAY_INTERSECT(x, y)",
+            },
+        )
+
+        self.validate_identity("SELECT ARRAY_INTERSECT(x, y, z)")
+
+        self.validate_all(
+            "ARRAY_REVERSE(x)",
+            read={
+                "clickhouse": "arrayReverse(x)",
+                "bigquery": "ARRAY_REVERSE(x)",
+                "snowflake": "ARRAY_REVERSE(x)",
+                "duckdb": "ARRAY_REVERSE(x)",
+            },
+            write={
+                "clickhouse": "arrayReverse(x)",
+                "bigquery": "ARRAY_REVERSE(x)",
+                "snowflake": "ARRAY_REVERSE(x)",
+                "duckdb": "ARRAY_REVERSE(x)",
+            },
+        )
+
+        self.validate_all(
+            "ARRAY_SLICE(x, 1, 3)",
+            read={
+                "clickhouse": "arraySlice(x, 1, 3)",
+                "bigquery": "ARRAY_SLICE(x, 1, 3)",
+                "snowflake": "ARRAY_SLICE(x, 1, 3)",
+                "duckdb": "ARRAY_SLICE(x, 1, 3)",
+                "spark2": "SLICE(x, 1, 3)",
+                "spark": "SLICE(x, 1, 3)",
+                "databricks": "SLICE(x, 1, 3)",
+                "presto": "SLICE(x, 1, 3)",
+                "trino": "SLICE(x, 1, 3)",
+            },
+            write={
+                "clickhouse": "arraySlice(x, 1, 3)",
+                "bigquery": "ARRAY_SLICE(x, 1, 3)",
+                "snowflake": "ARRAY_SLICE(x, 1, 3)",
+                "duckdb": "ARRAY_SLICE(x, 1, 3)",
+                "spark2": "SLICE(x, 1, 3)",
+                "spark": "SLICE(x, 1, 3)",
+                "databricks": "SLICE(x, 1, 3)",
+                "presto": "SLICE(x, 1, 3)",
+                "trino": "SLICE(x, 1, 3)",
+            },
+        )
+
     def test_order_by(self):
         self.validate_identity(
             "SELECT c FROM t ORDER BY a, b,",
@@ -1334,7 +1437,7 @@ class TestDialect(Validator):
                 "postgres": "JSON_EXTRACT_PATH(x, 'a b')",
                 "presto": """JSON_EXTRACT(x, '$["a b"]')""",
                 "redshift": "JSON_EXTRACT_PATH_TEXT(x, 'a b')",
-                "snowflake": """GET_PATH(x, '["a b"]')""",
+                "snowflake": """GET_PATH(PARSE_JSON(x), '["a b"]')""",
                 "spark": """GET_JSON_OBJECT(x, '$[\\'a b\\']')""",
                 "sqlite": """x -> '$."a b"'""",
                 "trino": """JSON_EXTRACT(x, '$["a b"]')""",
@@ -1363,7 +1466,7 @@ class TestDialect(Validator):
                 "oracle": "JSON_EXTRACT(x, '$.y')",
                 "postgres": "JSON_EXTRACT_PATH(x, 'y')",
                 "presto": "JSON_EXTRACT(x, '$.y')",
-                "snowflake": "GET_PATH(x, 'y')",
+                "snowflake": "GET_PATH(PARSE_JSON(x), 'y')",
                 "spark": "GET_JSON_OBJECT(x, '$.y')",
                 "sqlite": "x -> '$.y'",
                 "starrocks": "x -> '$.y'",
@@ -1418,7 +1521,7 @@ class TestDialect(Validator):
                 "postgres": "JSON_EXTRACT_PATH(x, 'y', '0', 'z')",
                 "presto": "JSON_EXTRACT(x, '$.y[0].z')",
                 "redshift": "JSON_EXTRACT_PATH_TEXT(x, 'y', '0', 'z')",
-                "snowflake": "GET_PATH(x, 'y[0].z')",
+                "snowflake": "GET_PATH(PARSE_JSON(x), 'y[0].z')",
                 "spark": "GET_JSON_OBJECT(x, '$.y[0].z')",
                 "sqlite": "x -> '$.y[0].z'",
                 "starrocks": "x -> '$.y[0].z'",
@@ -1472,7 +1575,7 @@ class TestDialect(Validator):
                 "clickhouse": "JSONExtractString(x, 'y')",
                 "postgres": "JSON_EXTRACT_PATH(x, 'y')",
                 "redshift": "JSON_EXTRACT_PATH_TEXT(x, 'y')",
-                "snowflake": "GET_PATH(x, 'y')",
+                "snowflake": "GET_PATH(PARSE_JSON(x), 'y')",
                 "sqlite": "x -> '$.y'",
                 "tsql": "ISNULL(JSON_QUERY(x, '$.y'), JSON_VALUE(x, '$.y'))",
             },
@@ -1522,7 +1625,9 @@ class TestDialect(Validator):
             "SELECT a FROM x CROSS JOIN UNNEST(y) WITH ORDINALITY AS t (a)",
             write={
                 "presto": "SELECT a FROM x CROSS JOIN UNNEST(y) WITH ORDINALITY AS t(a)",
-                "spark": "SELECT a FROM x LATERAL VIEW POSEXPLODE(y) t AS a",
+                "spark2": "SELECT a FROM x LATERAL VIEW POSEXPLODE(y) t AS pos, a",
+                "spark": "SELECT a FROM x LATERAL VIEW POSEXPLODE(y) t AS pos, a",
+                "databricks": "SELECT a FROM x LATERAL VIEW POSEXPLODE(y) t AS pos, a",
             },
         )
 
@@ -2185,6 +2290,21 @@ class TestDialect(Validator):
                 "bigquery": "MOD(a, b + 1)",
             },
         )
+        self.validate_all(
+            "ARRAY_REMOVE(the_array, target)",
+            write={
+                "": "ARRAY_REMOVE(the_array, target)",
+                "clickhouse": "arrayFilter(_u -> _u <> target, the_array)",
+                "duckdb": "LIST_FILTER(the_array, _u -> _u <> target)",
+                "bigquery": "ARRAY(SELECT _u FROM UNNEST(the_array) AS _u WHERE _u <> target)",
+                "hive": "ARRAY_REMOVE(the_array, target)",
+                "postgres": "ARRAY_REMOVE(the_array, target)",
+                "presto": "ARRAY_REMOVE(the_array, target)",
+                "starrocks": "ARRAY_REMOVE(the_array, target)",
+                "databricks": "ARRAY_REMOVE(the_array, target)",
+                "snowflake": "ARRAY_REMOVE(the_array, target)",
+            },
+        )
 
     def test_typeddiv(self):
         typed_div = exp.Div(this=exp.column("a"), expression=exp.column("b"), typed=True)
@@ -2619,6 +2739,7 @@ SELECT
                         "bigquery": f"LOG{base}(a)",
                         "clickhouse": f"LOG{base}(a)",
                         "databricks": f"LOG{base}(a)",
+                        "dremio": f"LOG{base}(a)",
                         "duckdb": f"LOG{base}(a)",
                         "mysql": f"LOG{base}(a)",
                         "postgres": f"LOG{base}(a)",
@@ -2631,6 +2752,7 @@ SELECT
                     write={
                         "bigquery": f"LOG(a, {base})",
                         "clickhouse": f"LOG{base}(a)",
+                        "dremio": f"LOG({base}, a)",
                         "duckdb": f"LOG({base}, a)",
                         "mysql": f"LOG({base}, a)",
                         "oracle": f"LOG({base}, a)",
@@ -2661,6 +2783,7 @@ SELECT
         self.validate_all(
             "LN(x)",
             read={
+                "dremio": "LOG(x)",
                 "bigquery": "LOG(x)",
                 "clickhouse": "LOG(x)",
                 "databricks": "LOG(x)",
@@ -2973,7 +3096,7 @@ FROM subquery2""",
                 "duckdb": "RANDOM()",
                 "hive": "RAND()",
                 "mysql": "RAND()",
-                "oracle": "RAND()",
+                "oracle": "DBMS_RANDOM.VALUE()",
                 "postgres": "RANDOM()",
                 "presto": "RAND()",
                 "spark": "RAND()",
@@ -2989,7 +3112,7 @@ FROM subquery2""",
                 "duckdb": "RANDOM()",
                 "hive": "RAND()",
                 "mysql": "RAND()",
-                "oracle": "RAND()",
+                "oracle": "DBMS_RANDOM.VALUE()",
                 "postgres": "RANDOM()",
                 "presto": "RAND()",
                 "spark": "RAND()",
@@ -3118,12 +3241,12 @@ FROM subquery2""",
             write={
                 "bigquery": "SELECT * FROM UNNEST(GENERATE_DATE_ARRAY(CAST('2020-01-01' AS DATE), CAST('2020-02-01' AS DATE), INTERVAL '1' WEEK))",
                 "databricks": "SELECT * FROM EXPLODE(SEQUENCE(CAST('2020-01-01' AS DATE), CAST('2020-02-01' AS DATE), INTERVAL '1' WEEK))",
-                "duckdb": "SELECT * FROM UNNEST(CAST(GENERATE_SERIES(CAST('2020-01-01' AS DATE), CAST('2020-02-01' AS DATE), (7 * INTERVAL '1' DAY)) AS DATE[]))",
+                "duckdb": "SELECT * FROM UNNEST(CAST(GENERATE_SERIES(CAST('2020-01-01' AS DATE), CAST('2020-02-01' AS DATE), INTERVAL '1' WEEK) AS DATE[]))",
                 "mysql": "WITH RECURSIVE _generated_dates(date_value) AS (SELECT CAST('2020-01-01' AS DATE) AS date_value UNION ALL SELECT CAST(DATE_ADD(date_value, INTERVAL 1 WEEK) AS DATE) FROM _generated_dates WHERE CAST(DATE_ADD(date_value, INTERVAL 1 WEEK) AS DATE) <= CAST('2020-02-01' AS DATE)) SELECT * FROM (SELECT date_value FROM _generated_dates) AS _generated_dates",
                 "postgres": "SELECT * FROM (SELECT CAST(value AS DATE) FROM GENERATE_SERIES(CAST('2020-01-01' AS DATE), CAST('2020-02-01' AS DATE), INTERVAL '1 WEEK') AS _t(value)) AS _unnested_generate_series",
                 "presto": "SELECT * FROM UNNEST(SEQUENCE(CAST('2020-01-01' AS DATE), CAST('2020-02-01' AS DATE), (1 * INTERVAL '7' DAY)))",
                 "redshift": "WITH RECURSIVE _generated_dates(date_value) AS (SELECT CAST('2020-01-01' AS DATE) AS date_value UNION ALL SELECT CAST(DATEADD(WEEK, 1, date_value) AS DATE) FROM _generated_dates WHERE CAST(DATEADD(WEEK, 1, date_value) AS DATE) <= CAST('2020-02-01' AS DATE)) SELECT * FROM (SELECT date_value FROM _generated_dates) AS _generated_dates",
-                "snowflake": "SELECT * FROM (SELECT DATEADD(WEEK, CAST(value AS INT), CAST('2020-01-01' AS DATE)) AS value FROM TABLE(FLATTEN(INPUT => ARRAY_GENERATE_RANGE(0, (DATEDIFF(WEEK, CAST('2020-01-01' AS DATE), CAST('2020-02-01' AS DATE)) + 1 - 1) + 1))) AS _u(seq, key, path, index, value, this))",
+                "snowflake": "SELECT * FROM (SELECT DATEADD(WEEK, CAST(value AS INT), CAST('2020-01-01' AS DATE)) AS value FROM TABLE(FLATTEN(INPUT => ARRAY_GENERATE_RANGE(0, (DATEDIFF(WEEK, CAST('2020-01-01' AS DATE), CAST('2020-02-01' AS DATE)) + 1 - 1) + 1))) AS _t0(seq, key, path, index, value, this))",
                 "spark": "SELECT * FROM EXPLODE(SEQUENCE(CAST('2020-01-01' AS DATE), CAST('2020-02-01' AS DATE), INTERVAL '1' WEEK))",
                 "trino": "SELECT * FROM UNNEST(SEQUENCE(CAST('2020-01-01' AS DATE), CAST('2020-02-01' AS DATE), (1 * INTERVAL '7' DAY)))",
                 "tsql": "WITH _generated_dates(date_value) AS (SELECT CAST('2020-01-01' AS DATE) AS date_value UNION ALL SELECT CAST(DATEADD(WEEK, 1, date_value) AS DATE) FROM _generated_dates WHERE CAST(DATEADD(WEEK, 1, date_value) AS DATE) <= CAST('2020-02-01' AS DATE)) SELECT * FROM (SELECT date_value AS date_value FROM _generated_dates) AS _generated_dates",
@@ -3158,7 +3281,7 @@ FROM subquery2""",
         self.validate_all(
             "SELECT ARRAY_LENGTH(GENERATE_DATE_ARRAY(DATE '2020-01-01', DATE '2020-02-01', INTERVAL 1 WEEK))",
             write={
-                "snowflake": "SELECT ARRAY_SIZE((SELECT ARRAY_AGG(*) FROM (SELECT DATEADD(WEEK, CAST(value AS INT), CAST('2020-01-01' AS DATE)) AS value FROM TABLE(FLATTEN(INPUT => ARRAY_GENERATE_RANGE(0, (DATEDIFF(WEEK, CAST('2020-01-01' AS DATE), CAST('2020-02-01' AS DATE)) + 1 - 1) + 1))) AS _u(seq, key, path, index, value, this))))",
+                "snowflake": "SELECT ARRAY_SIZE((SELECT ARRAY_AGG(*) FROM (SELECT DATEADD(WEEK, CAST(value AS INT), CAST('2020-01-01' AS DATE)) AS value FROM TABLE(FLATTEN(INPUT => ARRAY_GENERATE_RANGE(0, (DATEDIFF(WEEK, CAST('2020-01-01' AS DATE), CAST('2020-02-01' AS DATE)) + 1 - 1) + 1))) AS _t0(seq, key, path, index, value, this))))",
             },
         )
 
@@ -3463,3 +3586,189 @@ FROM subquery2""",
                     parse_one("SELECT 0xCC", read=read_dialect).sql(other_integer_dialects),
                     "SELECT 0xCC",
                 )
+
+    def test_ascii(self):
+        self.validate_all(
+            "ASCII('A')",
+            read={
+                "bigquery": "ASCII('A')",
+                "clickhouse": "ASCII('A')",
+                "databricks": "ASCII('A')",
+                "hive": "ASCII('A')",
+                "mysql": "ASCII('A')",
+                "postgres": "ASCII('A')",
+                "redshift": "ASCII('A')",
+                "snowflake": "ASCII('A')",
+                "tsql": "ASCII('A')",
+            },
+            write={
+                "bigquery": "ASCII('A')",
+                "clickhouse": "ASCII('A')",
+                "databricks": "ASCII('A')",
+                "hive": "ASCII('A')",
+                "mysql": "ASCII('A')",
+                "postgres": "ASCII('A')",
+                "redshift": "ASCII('A')",
+                "snowflake": "ASCII('A')",
+                "tsql": "ASCII('A')",
+            },
+        )
+
+    def test_bit_aggs(self):
+        self.validate_all(
+            "BIT_AND(x)",
+            read={
+                "bigquery": "BIT_AND(x)",
+                "spark": "BIT_AND(x)",
+                "databricks": "BIT_AND(x)",
+                "mysql": "BIT_AND(x)",
+                "dremio": "BIT_AND(x)",
+            },
+            write={
+                "bigquery": "BIT_AND(x)",
+                "spark": "BIT_AND(x)",
+                "databricks": "BIT_AND(x)",
+                "mysql": "BIT_AND(x)",
+                "dremio": "BIT_AND(x)",
+            },
+        )
+        self.validate_all(
+            "BIT_OR(x)",
+            read={
+                "bigquery": "BIT_OR(x)",
+                "spark": "BIT_OR(x)",
+                "databricks": "BIT_OR(x)",
+                "mysql": "BIT_OR(x)",
+                "dremio": "BIT_OR(x)",
+            },
+            write={
+                "bigquery": "BIT_OR(x)",
+                "spark": "BIT_OR(x)",
+                "databricks": "BIT_OR(x)",
+                "mysql": "BIT_OR(x)",
+                "dremio": "BIT_OR(x)",
+            },
+        )
+        self.validate_all(
+            "BIT_XOR(x)",
+            read={
+                "bigquery": "BIT_XOR(x)",
+                "spark": "BIT_XOR(x)",
+                "databricks": "BIT_XOR(x)",
+                "mysql": "BIT_XOR(x)",
+            },
+            write={
+                "bigquery": "BIT_XOR(x)",
+                "spark": "BIT_XOR(x)",
+                "databricks": "BIT_XOR(x)",
+                "mysql": "BIT_XOR(x)",
+            },
+        )
+        self.validate_all(
+            "BIT_COUNT(x)",
+            read={
+                "bigquery": "BIT_COUNT(x)",
+                "spark": "BIT_COUNT(x)",
+                "databricks": "BIT_COUNT(x)",
+                "mysql": "BIT_COUNT(x)",
+            },
+            write={
+                "bigquery": "BIT_COUNT(x)",
+                "spark": "BIT_COUNT(x)",
+                "databricks": "BIT_COUNT(x)",
+                "mysql": "BIT_COUNT(x)",
+            },
+        )
+
+    def test_between(self):
+        self.validate_all(
+            "SELECT x BETWEEN 2 AND 10",
+            read={
+                "": "SELECT x BETWEEN 2 AND 10",
+                "clickhouse": "SELECT x BETWEEN 2 AND 10",
+                "dremio": "SELECT x BETWEEN 2 AND 10",
+                "duckdb": "SELECT x BETWEEN 2 AND 10",
+                "tsql": "SELECT x BETWEEN 2 AND 10",
+                "oracle": "SELECT x BETWEEN 2 AND 10",
+                "mysql": "SELECT x BETWEEN 2 AND 10",
+                "postgres": "SELECT x BETWEEN 2 AND 10",
+            },
+            write={
+                "": "SELECT x BETWEEN 2 AND 10",
+                "clickhouse": "SELECT x BETWEEN 2 AND 10",
+                "dremio": "SELECT x BETWEEN 2 AND 10",
+                "duckdb": "SELECT x BETWEEN 2 AND 10",
+                "tsql": "SELECT x BETWEEN 2 AND 10",
+                "oracle": "SELECT x BETWEEN 2 AND 10",
+                "mysql": "SELECT x BETWEEN 2 AND 10",
+                "postgres": "SELECT x BETWEEN 2 AND 10",
+            },
+        )
+
+        self.validate_all(
+            "SELECT x BETWEEN SYMMETRIC 10 AND 2",
+            write={
+                "dremio": "SELECT x BETWEEN SYMMETRIC 10 AND 2",
+                "postgres": "SELECT x BETWEEN SYMMETRIC 10 AND 2",
+                "": "SELECT (x BETWEEN 10 AND 2 OR x BETWEEN 2 AND 10)",
+                "mysql": "SELECT (x BETWEEN 10 AND 2 OR x BETWEEN 2 AND 10)",
+                "oracle": "SELECT (x BETWEEN 10 AND 2 OR x BETWEEN 2 AND 10)",
+                "duckdb": "SELECT (x BETWEEN 10 AND 2 OR x BETWEEN 2 AND 10)",
+                "clickhouse": "SELECT (x BETWEEN 10 AND 2 OR x BETWEEN 2 AND 10)",
+                "tsql": "SELECT (x BETWEEN 10 AND 2 OR x BETWEEN 2 AND 10)",
+            },
+        )
+
+        self.validate_all(
+            "SELECT x BETWEEN ASYMMETRIC 10 AND 2",
+            write={
+                "dremio": "SELECT x BETWEEN ASYMMETRIC 10 AND 2",
+                "postgres": "SELECT x BETWEEN ASYMMETRIC 10 AND 2",
+                "": "SELECT x BETWEEN 10 AND 2",
+                "mysql": "SELECT x BETWEEN 10 AND 2",
+                "oracle": "SELECT x BETWEEN 10 AND 2",
+                "duckdb": "SELECT x BETWEEN 10 AND 2",
+                "clickhouse": "SELECT x BETWEEN 10 AND 2",
+                "tsql": "SELECT x BETWEEN 10 AND 2",
+            },
+        )
+
+    def test_like_quantifiers(self):
+        for quantifier in ("ANY", "ALL"):
+            connector = "OR" if quantifier == "ANY" else "AND"
+
+            with self.subTest(f"Testing LIKE {quantifier}"):
+                self.validate_all(
+                    f"SELECT col LIKE {quantifier} (x, y, z)",
+                    read={
+                        "": f"SELECT col LIKE {quantifier} (x, y, z)",
+                        "bigquery": f"SELECT col LIKE {quantifier} (x, y, z)",
+                        "snowflake": f"SELECT col LIKE {quantifier} (x, y, z)",
+                        "spark": f"SELECT col LIKE {quantifier} (x, y, z)",
+                        "databricks": f"SELECT col LIKE {quantifier} (x, y, z)",
+                    },
+                    write={
+                        "bigquery": f"SELECT col LIKE {quantifier} (x, y, z)",
+                        "snowflake": f"SELECT col LIKE {quantifier} (x, y, z)",
+                        "spark": f"SELECT col LIKE {quantifier} (x, y, z)",
+                        "databricks": f"SELECT col LIKE {quantifier} (x, y, z)",
+                        "duckdb": f"SELECT (col LIKE x {connector} col LIKE y) {connector} col LIKE z",
+                    },
+                )
+
+            with self.subTest(f"Testing ILIKE {quantifier}"):
+                self.validate_all(
+                    f"SELECT col ILIKE {quantifier} (x, y, z)",
+                    write={
+                        "": f"SELECT col ILIKE {quantifier} (x, y, z)",
+                        "duckdb": f"SELECT (col ILIKE x {connector} col ILIKE y) {connector} col ILIKE z",
+                    },
+                )
+
+        self.validate_all(
+            "SELECT 'foo' LIKE ANY((('bar', 'fo%')))",
+            write={
+                "": "SELECT 'foo' LIKE ANY((('bar', 'fo%')))",
+                "duckdb": "SELECT 'foo' LIKE 'bar' OR 'foo' LIKE 'fo%'",
+            },
+        )
